tests/wpt/web-platform-tests/focus/activeelement-after-immediately-focusing-different-site-iframe-contentwindow.html

<!doctype html>
<meta charset=utf-8>
<title>activeElement when immediately focusing different-site iframe's contentWindow</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script>
setup({explicit_done:true});
window.onmessage = function(e) {
    var actual = e.data;
    test(function() {
        // Make the difference between Firefox and Chrome visible separately
        // from the comparison of the entire log string failing to match.
        var endedWith = false;
        if (actual.endsWith(",willspineventloop,")) {
            endedWith = true;
            actual += "innerbodyfocus,";
        }
        assert_true(endedWith, "Should not have gotten innerbodyfocus after willspineventloop");
    }, "Check trailing events");
    test(function() {
        assert_equals(actual, "outerparser,activeElement:BODY,willfocusiframe,didfocusiframe,activeElement:IFRAME,willbluriframe,didbluriframe,activeElement:IFRAME,willspineventloop,innerbodyfocus,", 'Check log');
    }, "Check result");
    w.close();
    done();
};
var w = window.open("support/activeelement-after-immediately-focusing-different-site-iframe-outer-contentwindow.sub.html");
</script>
